# docs/v1/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateQuestionRequest:
    properties:
      content:
        minLength: 10
        type: string
      tags:
        description: Comma-separated tags
        type: string
      title:
        maxLength: 255
        minLength: 5
        type: string
      user_id:
        type: string
    required:
    - content
    - title
    - user_id
    type: object
...
  dto.UpdateQuestionRequest:
    properties:
      content:
        minLength: 10
        type: string
      tags:
        type: string
      title:
        maxLength: 255
        minLength: 5
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      message:
        type: string
      meta:
        $ref: '#/definitions/response.MetaInfo'
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  response.ErrorInfo:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  response.MetaInfo:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: A RESTful API built with Go and Fiber
  title: Pitstop API
  version: "1.0"
paths:
...
  /questions:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of questions
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Questions per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get all questions
      tags:
      - questions
    post:
      consumes:
      - application/json
      description: Create a new question
      parameters:
      - description: Question details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateQuestionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new question
      tags:
      - questions
  /questions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a question (only by author)
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a question
      tags:
      - questions
    get:
      consumes:
      - application/json
      description: Retrieve a specific question
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get a question by ID
      tags:
      - questions
    put:
      consumes:
      - application/json
      description: Update a question (only by author)
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated question details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a question
      tags:
      - questions
...
  /questions/tag:
    get:
      consumes:
      - application/json
      description: Retrieve questions filtered by a specific tag
      parameters:
      - description: Tag to filter by
        in: query
        name: tag
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Questions per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get questions by tag
      tags:
      - questions
...
swagger: "2.0"